Add Notifications demo
This code does not appear to work. I successfully create the
notification (I stepped into the line in the debugger), but I see no
user-visible indication that I created the notification. Unfortunately,
I do not really know what to expect. For now, I'll check in the code and
move to another demo.